Use the developer logging facilities from the dart SDK
It is quite rude to override the debugPrint function. Generally I am not a fan
of such blanket solutions. The way to go is to use dart's logging tools.
They provide much more fine-grained control over the logs. These logs are not
enabled when the app is compiled in release mode.

https://docs.flutter.dev/testing/code-debugging#logging