Rework dio interceptor handling
Create and add interceptors when the app is initialized and not
when the app widget is rebuild.

Update the auth interceptor to not depend directly on dio.
Also do not log on all errors but only on auth errors.

Get rid of the half-finished custom log interceptor in favor of
the dio-provided one.